Move Xiaomi special reporting related code to dedicated file
